requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==24.1.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
boto3==1.38.28
    # via -r requirements.in
botocore==1.38.28
    # via
    #   boto3
    #   s3transfer
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via
    #   pdfminer-six
    #   requests
click==8.2.1
    # via uvicorn
cryptography==45.0.3
    # via pdfminer-six
distro==1.9.0
    # via openai
fastapi==0.115.9
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
grpcio==1.72.1
    # via qdrant-client
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hf-xet==1.1.2
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   openai
    #   qdrant-client
huggingface-hub==0.32.3
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via torch
jiter==0.10.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.1
    # via scikit-learn
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
mypy==1.16.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
networkx==3.5
    # via torch
numpy==2.2.6
    # via
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
openai==1.83.0
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   pytest
    #   transformers
pathspec==0.12.1
    # via mypy
pdfminer-six==20250327
    # via pdfplumber
pdfplumber==0.11.6
    # via -r requirements.in
pillow==11.2.1
    # via
    #   pdfplumber
    #   sentence-transformers
pip==25.1.1
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
portalocker==2.10.1
    # via qdrant-client
protobuf==5.29.5
    # via qdrant-client
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pygments==2.19.1
    # via pytest
pymupdf==1.26.0
    # via -r requirements.in
pypdfium2==4.30.1
    # via pdfplumber
pytest==8.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
qdrant-client==1.14.2
    # via -r requirements.in
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   tiktoken
    #   transformers
s3transfer==0.13.0
    # via boto3
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
    # via -r requirements.in
setuptools==80.9.0
    # via triton
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.45.3
    # via fastapi
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via -r requirements.in
tokenizers==0.21.1
    # via transformers
torch==2.7.0
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.52.4
    # via sentence-transformers
triton==3.3.0
    # via torch
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   botocore
    #   qdrant-client
    #   requests
uuid6==2024.7.10
    # via -r requirements.in
uvicorn==0.34.3
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
